// src/utils/enums/enums.ts
export enum TabsEnum {
  MEAL_CALCULATOR = 'Meal Calculator',
  PROTEIN_PER_SEK = 'Protein Per Sek',
}

export enum UnitsEnum {
  GRAMS = 'gram',
  KILOGRAMS = 'kilo',
  LITERS = 'liter',
  UNIT = 'styck',
  DECILITERS = 'deciliter',
  MILLILITERS = 'milliliter',
}

export enum ComparisonPriceUnitsEnum {
  KILOGRAMS = 'kr/kg',
  LITERS = 'kr/l',
  UNIT = 'kr/st',
}